реализвоал полный цикл: LocationController -> LocationService-> Client or JpaLocationRepository

+++ src/main/java/com/example/weatherviewer/client/LocationResponse.java
@@ -1,9 +1,13 @@
 package com.example.weatherviewer.client;
 
+
+import java.math.BigDecimal;
+
+
 public record LocationResponse(
         String name,
-        double latitude,
-        double longitude,
+        BigDecimal latitude,
+        BigDecimal longitude,
         String country,
         String state
 ) {
